split functions to a single action functions
