+++ js/src/functions/itemContsructor.js
@@ -1,0 +1,7 @@
+export function Item(name, sell_in, quality){
+  return {
+    name,
+    sell_in,
+    quality,
+  }
+}
